tests-e2e/requirements.txt
pytest==8.3.3
pytest-xdist==3.6.1
requests==2.32.3
psycopg2-binary==2.9.10
tenacity==9.0.0
PyYAML==6.0.2
pytest-html==4.1.1
